move pages stories to separate folder
